+++ examples/next-drizzle-mysql/drizzle/0000_brave_madrox.sql
@@ -1,0 +1,6 @@
+CREATE TABLE `users` (
+	`id` int AUTO_INCREMENT NOT NULL,
+	`name` json,
+	`email` json,
+	CONSTRAINT `users_id` PRIMARY KEY(`id`)
+);