+++ examples/next-drizzle-mysql/src/protect/index.ts
@@ -1,0 +1,4 @@
+import { protect } from '@cipherstash/protect'
+import { users } from './schema'
+
+export const protectClient = await protect(users)