+++ examples/basic/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @cipherstash/basic-example
 
+## 1.0.12
+
+### Patch Changes
+
+- Updated dependencies [a471821]
+  - @cipherstash/protect@8.4.0
+
+## 1.0.11
+
+### Patch Changes
+
+- Updated dependencies [628acdc]
+  - @cipherstash/protect@8.3.0
+
 ## 1.0.10
 
 ### Patch Changes