+++ packages/protect/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @cipherstash/jseql
 
+## 8.4.0
+
+### Minor Changes
+
+- a471821: Fixed a bug in the model interface to correctly handle undefined and null values.
+
+## 8.3.0
+
+### Minor Changes
+
+- 628acdc: Implemented createSearchTerms for a streamlined way of working with encrypted search terms.
+
 ## 8.2.0
 
 ### Minor Changes